+++ src/components/ComponentRegistry.ts
@@ -12,6 +12,7 @@ import {Low} from "@/components/input/Low.tsx";
 import {Light} from "@/components/output/Light.tsx";
 import {Multiplexer} from "@/components/multiplexer/Multiplexer.tsx";
 import {Demultiplexer} from "@/components/multiplexer/Demultiplexer.tsx";
+import {latches} from "@/components/latches";
 
 export const componentRegistry = {
     // Gates
@@ -32,6 +33,8 @@ export const componentRegistry = {
     // Multiplexer
     multiplexer: Multiplexer,
     demultiplexer: Demultiplexer,
+    // Latches
+    ...latches
 }
 
 export type ComponentType = keyof typeof componentRegistry;