+++ src/components/ComponentTree.tsx
@@ -15,11 +15,16 @@ import NORGateSVG from '../assets/components/gates/NOrGate.svg?react';
 import XORGateSVG from '../assets/components/gates/XOrGate.svg?react';
 import XNORGateSVG from '../assets/components/gates/XNOrGate.svg?react';
 import MultiplexerSVG from '../assets/components/multiplexer/Multiplexer.svg?react';
+import SRLatchSVG from '@assets/components/latches/SRLatch.svg?react';
+import DFlipFlopSVG from '@assets/components/latches/DFlipFlop.svg?react';
+import JKFlipFlopSVG from '@assets/components/latches/JKFlipFlop.svg?react';
+import TFlipFlopSVG from '@assets/components/latches/TFlipFlop.svg?react';
 import './ComponentTree.css';
 import {LogicGate} from "./gates";
 import {Output} from "@/components/output";
 import {Input} from "@/components/input";
 import {MultiplexerType} from "@/components/multiplexer";
+import {Latch} from "@/components/latches";
 
 export default function ComponentTree() {
     const {setNodes, screenToFlowPosition} = useReactFlow();
@@ -75,6 +80,12 @@ export default function ComponentTree() {
             <ComponentCategory label={"Multiplexers"}>
                 <ComponentItem componentID={MultiplexerType.MUX} label={"MUX"} icon={MultiplexerSVG} onDrop={handleNodeDrop}/>
                 <ComponentItem componentID={MultiplexerType.DMUX} label={"DMUX"} icon={MultiplexerSVG} flip onDrop={handleNodeDrop}/>
+            </ComponentCategory>
+            <ComponentCategory label={"Latches"}>
+                <ComponentItem componentID={Latch.SR_LATCH} label={"SR Latch"} icon={SRLatchSVG} onDrop={handleNodeDrop}/>
+                <ComponentItem componentID={Latch.D_FLIP_FLOP} label={"D flip-flop"} icon={DFlipFlopSVG} onDrop={handleNodeDrop}/>
+                <ComponentItem componentID={Latch.JK_FLIP_FLOP} label={"JK flip-flop"} icon={JKFlipFlopSVG} onDrop={handleNodeDrop}/>
+                <ComponentItem componentID={Latch.T_FLIP_FLOP} label={"T flip-flop"} icon={TFlipFlopSVG} onDrop={handleNodeDrop}/>
             </ComponentCategory>
         </div>
     )