+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Target handle ID in evaluation function
+- SR Latch
+- D, T, and JK Flip-Flops
+
 ## [0.1.0]
 
 ### Added